requirements.txt
streamlit
plotly
pandas
numpy
websockets
cryptography
python-dotenv
requests
openai
